Drawing code for grids
Grids are drawn using triangle strips in GL. You can render a whole
tessalated grid in a single draw call if you're careful with your
indices. This takes the approach of rendering a row of right angle
triangles then a mirrored row of right angle triangles below it, this is
required to be able to traverse everything in a single draw call.

At the end of each row, three degenerate triangles must be rendered to
keep the winding order consistent.

See
http://dan.lecocq.us/wordpress/2009/12/25/triangle-strip-for-grids-a-construction/
for a better explanation.